feat(viewer): align slider ranges/defaults with prototype spec
Speed: 1-10 → 0.5-3.0 (step 0.1, default 1.0x)
FontSize: 20-100 → 16-64 (default 32px)
Opacity min: 0 → 20 (default 100%)

- slider value attrs now use options[3] (real default) not midpoint
- viewer-range: appearance:auto restores native track so accent-color works
- overlay.js: parseFloat + updated duration formula + clamp to 0.5-3.0
- admin.js: speed inputs step="0.1"
- all 881 unit tests updated and passing

+++ server/static/js/main.js
@@ -767,7 +767,7 @@ document.addEventListener("DOMContentLoaded", () => {
         text: text,
         color: elements.colorInput ? elements.colorInput.value : null,
         size: elements.sizeInput ? parseInt(elements.sizeInput.value) : null,
-        speed: elements.speedRange ? parseInt(elements.speedRange.value) : null,
+        speed: elements.speedRange ? parseFloat(elements.speedRange.value) : null,
         opacity: elements.opacityRange
           ? parseInt(elements.opacityRange.value)
           : null,
